api/update_event/requirements.txt
openai
psycopg2-binary
python-dotenv
